requirements.txt
opencv-python==4.6.0.66
paddlepaddle==2.5.1
paddleocr==2.7.0.3
pywin32==306
numpy==1.24.3
pillow==10.0.0
pyautogui==0.9.53
